Tested previous PR and made changes
@dcummings95
Changed:
Global.gd -> Player.gd
- This is to keep a more consistent naming scheme since everything in there is related to player.

Game.gd
- Moved these to to Player.gd. We can circle back on the purpose of this one OR we can change this to be more upgrade oriented which the name will have to change.

Created UI Footer for Level1 for testing. I am unsure if we should do a specific scene for each level (we can), but this can be farther discussed.

Made the HP/Shields/Stamina to work with Player.gd along with taking damage to apply on shields first if shield is not 0. Additionally, Stamina is on a timer which how I have it now it will take 5 seconds to fill from 0 - 100. Not sure if Stamina is where we want it to be, but that is balancing for the future.

Also, I created a directory for Scripts. We will separate scenes and Scripts to easily find them for the future.

Also, I need to know what you changed with the Grid. This is important since it kind of ruins the algorithm I have created for Grid Spawning for World/Event type Grids.